jina/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiostream==0.4.5
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
Deprecated==1.2.13
docarray==0.19.1
docker==6.0.1
fastapi==0.87.0
filelock==3.8.0
flatbuffers==22.11.23
fonttools==4.38.0
frozenlist==1.3.3
googleapis-common-protos==1.57.0
grpcio==1.47.2
grpcio-health-checking==1.47.2
grpcio-reflection==1.47.2
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.1.0
jc==1.22.2
jcloud==0.1.2
jina==3.11.2
jina-hubble-sdk==0.26.4
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.6.2
mpmath==1.2.1
multidict==6.0.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnxruntime==1.13.1
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp==1.14.0
opentelemetry-exporter-otlp-proto-grpc==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-instrumentation-fastapi==0.35b0
opentelemetry-instrumentation-grpc==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==21.3
pathspec==0.10.2
Pillow==9.3.0
prometheus-client==0.15.0
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.8.1
zipp==3.10.0
